requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
yarl==1.9.4
flake8==7.0.0
flake8-isort==6.1.1